ESCALONADOR QUASE FUNCIONANDO
falta só conectar o pid do processo com o pid lido pelo escalonador e tratar concorrência
