--- queue.h
@@ -1,4 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-#include "info.h"
