+++ src/RealTime/Makefile
@@ -1,6 +1,7 @@
 all:
+	clear
 	gcc -Wall -o interpretador_RT interpretador_RT.c
-	gcc -Wall -o escalonador_RT escalonador_RT.c queue.h
+	gcc -Wall -o escalonador_RT queue.c escalonador_RT.c 
 	./interpretador_RT
 
 interpretar: interpretador_RT.c
