+++ src/RealTime/exec.txt
@@ -1,5 +1,5 @@
-Run <P1> I=<55> D=<2>
-Run <P2> I=<50> D=<3>
-Run <P3> I=<40> D=<5>
-Run <P4> I=<30> D=<5>
-Run <P5> I=<10> D=<10>
+Run <P1> I=<10> D=<5>
+Run <P2> I=<20> D=<5>
+Run <P3> I=<30> D=<5>
+Run <P4> I=<40> D=<5>
+Run <P5> I=<50> D=<5>
