+++ src/programas/Makefile
@@ -1,0 +1,7 @@
+all:
+	clear
+	gcc -Wall -o P1 P1.c
+	gcc -Wall -o P2 P2.c
+	gcc -Wall -o P3 P3.c
+	gcc -Wall -o P4 P4.c
+	gcc -Wall -o P5 P5.c
